Switched file_path method to with_file_path, which yields the right file path.
In cases where the file has not been copied to its final location yet, the
with_file_path method copies the file to a a temp file with the file's correct
filename, which lets scribd handle it correctly. It then ensures the deletion of
the file when the block ends.